+++ src/components/WatchlistDisplay.css
@@ -1,0 +1,8 @@
+.watchlist {
+  border: 2px solid black;
+  margin-left: 5vw;
+  margin-right: 5vw;
+  justify-self: center;
+  display: flex;
+  flex-flow: column;
+}
